requirements.txt
nb-cli==0.6.8
nonebot-adapter-onebot==2.1.5
numpy
pandas
nonebot2==2.0.0rc2